typings, refactor changes: * refactor `ZookeeperLock` to extend `EventEmitter` instead of wrapping * use typings instead of tsd * use tsconfig.js instead of grunt to compile * bluebird instead of q * update locators
