+++ gruntfile.coffee
@@ -4,25 +4,12 @@ module.exports = (grunt) ->
   grunt.loadNpmTasks('grunt-contrib-watch')
   grunt.loadNpmTasks('grunt-ts')
   grunt.loadNpmTasks('grunt-tslint')
-  grunt.loadNpmTasks('grunt-tsd')
   grunt.loadNpmTasks('dts-generator')
 
   config =
     ts:
-      options:
-        target: 'es5'
-        module: 'commonjs'
-        declaration: true
-        sourceMap: true
-        emitDecoratorMetadata: true
-        failOnTypeErrors: false
       build:
-        src:  [ 'src/*.ts', 'typings/**/*.ts', 'typings_custom/*.ts' ]
-        outDir: 'build/'
-        baseDir: 'src/'
-        options:
-          sourceMap: true
-          noEmitHelpers: true
+        tsconfig: 'tsconfig.json'
 
     tslint:
       options:
@@ -30,28 +17,14 @@ module.exports = (grunt) ->
       files:
         src: [ 'src/**/*.ts' ]
 
-    tsd:
-      load:
-        options:
-          command: 'reinstall'
-          latest: false
-          config: 'tsd.json'
-      refresh:
-        options:
-          command: 'reinstall'
-          latest: true
-          config: 'tsd.json'
-
     dtsGenerator:
       options:
         name: 'zk-lock'
         baseDir: 'src/'
         out: 'build/zk-lock.d.ts'
-        main: 'zk-lock/zookeeperLock'
+        main: 'zk-lock/index'
       default:
         src: [ 'src/**/*.ts' ]
-
-
 
     watch:
       typescripts:
@@ -64,7 +37,6 @@ module.exports = (grunt) ->
   grunt.initConfig(config)
 
   grunt.registerTask 'default', [
-    'tsd:load',
     'tslint',
     'ts:build',
     'dtsGenerator'
