+++ package.json
@@ -2,34 +2,47 @@
   "name": "zk-lock",
   "version": "0.0.17",
   "description": "A distributed lock using zookeeper",
+  "keywords": [
+    "zookeeper",
+    "zookeeper lock",
+    "distributed lock"
+  ],
   "license": "ISC",
+  "repository": {
+    "type": "git",
+    "url": "http://github.com/metamx/zk-beacon.git"
+  },
+  "main": "build/index.js",
+  "typescript": {
+    "definition": "build/zk-lock.d.ts"
+  },
+  "scripts": {
+    "typings": "node_modules/.bin/typings install",
+    "compile": "node_modules/.bin/grunt",
+    "build": "npm run typings && npm run compile",
+    "test": "node_modules/.bin/mocha",
+    "debugtest": "NODE_DEBUG=zk-lock ./node_modules/.bin/mocha",
+    "prepublish": "npm run build"
+  },
   "dependencies": {
-    "locators": "^0.3.8",
-    "q": "^1.4.1",
+    "bluebird": "^3.4.2",
+    "locators": "^1.0.0",
     "node-zookeeper-client": "^0.2.2"
   },
   "devDependencies": {
-    "chai": "^3.4.1",
+    "chai": "^3.5.0",
     "coffee-script": "^1.10.0",
-    "dts-generator": "^1.6.3",
-    "grunt": "^0.4.5",
-    "grunt-cli": "^0.1.13",
-    "grunt-contrib-copy": "^0.8.2",
-    "grunt-contrib-watch": "^0.6.1",
-    "grunt-ts": "^5.2.0",
-    "grunt-tsd": "^0.2.0-beta.1",
-    "grunt-tslint": "^3.0.0",
-    "mocha": "^2.3.4",
-    "nock": "^3.3.2",
-    "tsd": "^0.6.5",
-    "tslint": "^3.1.1",
-    "typescript": "^1.7.3"
-  },
-  "scripts": {
-    "build": "node_modules/.bin/grunt",
-    "test": "node_modules/.bin/mocha",
-    "debugtest": "NODE_DEBUG=zk-lock ./node_modules/.bin/mocha",
-    "prepublish": "node_modules/.bin/grunt"
-  },
-  "main": "build/zookeeperLock.js"
+    "dts-generator": "^1.7.0",
+    "grunt": "^1.0.1",
+    "grunt-cli": "^1.2.0",
+    "grunt-contrib-copy": "^1.0.0",
+    "grunt-contrib-watch": "^1.0.0",
+    "grunt-ts": "^5.5.1",
+    "grunt-tslint": "^3.2.1",
+    "mocha": "^3.0.2",
+    "nock": "^8.0.0",
+    "tslint": "^3.15.1",
+    "typescript": "^1.8.10",
+    "typings": "^1.3.2"
+  }
 }
