+++ tslint.json
@@ -25,8 +25,7 @@
     "no-eval": true,
     "no-shadowed-variable": true,
     "no-string-literal": false,
-    "no-switch-case-fall-through": true,
-    "no-trailing-comma": true,
+    "no-switch-case-fall-through": false,
     "no-trailing-whitespace": false,
     "no-unused-expression": true,
     "no-unused-variable": false,
@@ -40,6 +39,7 @@
     ],
     "radix": true,
     "semicolon": true,
+    "trailing-comma": true,
     "triple-equals": [true, "allow-null-check"],
     "variable-name": false,
     "whitespace": [true,
