+++ typings_local/locators.d.ts
